fix(lint): extract auto model resolution to stay under chatHelpers LOC freeze
